functions/paragraph-ocr/image-to-text-0/requirements.txt
minio
pillow
pytesseract
